Rename README.md to PhamTrungThanh.md
Hello, my name is Thanh An. I’m 15 years old and I’m in grade 9. I was born and grew up in HCM city. I feel really honored for being chosen to be the best student of our class in this semester. I promise to study harder and harder to accomplish your believe and also help others in studying.